Make the RedHat RPMs build again after the r8778 change.

Approved by: david, kfogel


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/1.0.x@848860 13f79535-47bb-0310-9956-ffa450edef68

+++ packages/rpm/redhat-8+/subversion.spec
@@ -105,6 +105,11 @@ Summary: Tools for Subversion
 Tools for Subversion.
 
 %changelog
+* Thu Feb 19 2004 David Summers <david@summersoft.fay.ar.us> 0.37.0-8770
+- cvs2svn.1 man page taken out of distribution.  Change RPM so that if it
+  just so happens to magically reappear someday then it will get put into
+  the package.
+
 * Sun Jan 18 2004 David Summers <david@summersoft.fay.ar.us> 0.36.0-8372
 - Switched to the Redhat way of doing the "swig" package where it is not
   separated into "swig" and "swig-runtime".
@@ -424,7 +429,9 @@ mkdir -p $RPM_BUILD_ROOT/usr/lib/python2.2/site-packages
 cp -r tools/cvs2svn/rcsparse $RPM_BUILD_ROOT/usr/lib/python2.2/site-packages/rcsparse
 mv $RPM_BUILD_ROOT/usr/lib/svn-python/* $RPM_BUILD_ROOT/usr/lib/python2.2/site-packages
 rmdir $RPM_BUILD_ROOT/usr/lib/svn-python
-cp $RPM_BUILD_DIR/subversion-%{version}/tools/cvs2svn/cvs2svn.1 $RPM_BUILD_ROOT/usr/share/man/man1
+if [ -f $RPM_BUILD_DIR/subversion-%{version}/tools/cvs2svn/cvs2svn.1 ]; then
+   cp $RPM_BUILD_DIR/subversion-%{version}/tools/cvs2svn/cvs2svn.1 $RPM_BUILD_ROOT/usr/share/man/man1
+fi
 
 # Install PERL SWIG bindings.
 make install-swig-pl-lib DESTDIR=$RPM_BUILD_ROOT
